Adds basic test for pushing content + removes getArtworks
More context: Luc and I were hoping to be able to use ElasticSearch to get artwork and artist data associated with each collection. However, we realized we'd have to make big changes to make that happen, so we're going to query Metaphysics instead. Starting a separate PR for that work.

Co-authored-by: Eve Essex <eve.essex@gmail.com>